requirements.txt
sumy
newspaper3k
